chore: fixing PR suggestions
