+++ docs/supported-libraries.md
@@ -30,6 +30,7 @@ These are the supported libraries and frameworks:
 | [Apache Dubbo](https://github.com/apache/dubbo/)                                                                                            | 2.7+                          | [opentelemetry-apache-dubbo-2.7](../instrumentation/apache-dubbo-2.7/library-autoconfigure)                                                                                                                                                                                                                                                                                             | [RPC Client Spans], [RPC Server Spans]                                                                                                  |
 | [Apache HttpAsyncClient](https://hc.apache.org/index.html)                                                                                  | 4.1+                          | N/A                                                                                                                                                                                                                                                                                                                                                                                     | [HTTP Client Spans], [HTTP Client Metrics]                                                                                              |
 | [Apache HttpClient](https://hc.apache.org/index.html)                                                                                       | 2.0+                          | [opentelemetry-apache-httpclient-4.3](../instrumentation/apache-httpclient/apache-httpclient-4.3/library)                                                                                                                                                                                                                                                                               | [HTTP Client Spans], [HTTP Client Metrics]                                                                                              |
+| [Apache HttpClient 5](https://hc.apache.org/index.html)                                                                                     | 2.0+                          | [opentelemetry-apache-httpclient-5.2](../instrumentation/apache-httpclient/apache-httpclient-5.2/library)                                                                                                                                                                                                                                                                               | [HTTP Client Spans], [HTTP Client Metrics]                                                                                              |
 | [Apache Kafka Producer/Consumer API](https://kafka.apache.org/documentation/#producerapi)                                                   | 0.11+                         | [opentelemetry-kafka-clients-2.6](../instrumentation/kafka/kafka-clients/kafka-clients-2.6/library)                                                                                                                                                                                                                                                                                     | [Messaging Spans]                                                                                                                       |
 | [Apache Kafka Streams API](https://kafka.apache.org/documentation/streams/)                                                                 | 0.11+                         | N/A                                                                                                                                                                                                                                                                                                                                                                                     | [Messaging Spans]                                                                                                                       |
 | [Apache MyFaces](https://myfaces.apache.org/)                                                                                               | 1.2+ (not including 3.x yet)  | N/A                                                                                                                                                                                                                                                                                                                                                                                     | Provides `http.route` [2], Controller Spans [3]                                                                                         |
