--- instrumentation/apache-httpclient/apache-httpclient-5.0/library/src/main/java/io/opentelemetry/instrumentation/apachehttpclient/v5_0/HttpHeaderSetter.java
@@ -1,22 +1,0 @@
-/*
- * Copyright The OpenTelemetry Authors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-package io.opentelemetry.instrumentation.apachehttpclient.v5_0;
-
-import io.opentelemetry.context.propagation.TextMapSetter;
-import jakarta.annotation.Nullable;
-import org.apache.hc.core5.http.HttpRequest;
-
-enum HttpHeaderSetter implements TextMapSetter<HttpRequest> {
-  INSTANCE;
-
-  @Override
-  public void set(@Nullable HttpRequest carrier, String key, String value) {
-    if (carrier == null) {
-      return;
-    }
-    carrier.setHeader(key, value);
-  }
-}
