+++ instrumentation/apache-httpclient/apache-httpclient-5.2/library/build.gradle.kts
@@ -1,0 +1,10 @@
+plugins {
+  id("otel.library-instrumentation")
+  id("otel.nullaway-conventions")
+  id("otel.animalsniffer-conventions")
+}
+
+dependencies {
+  library("org.apache.httpcomponents.client5:httpclient5:5.2.1")
+  library("jakarta.annotation:jakarta.annotation-api:2.1.1")
+}
