+++ instrumentation/apache-httpclient/apache-httpclient-5.2/library/src/main/java/io/opentelemetry/instrumentation/apachehttpclient/v5_2/HttpHeaderSetter.java
@@ -1,0 +1,22 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package io.opentelemetry.instrumentation.apachehttpclient.v5_2;
+
+import io.opentelemetry.context.propagation.TextMapSetter;
+import jakarta.annotation.Nullable;
+import org.apache.hc.core5.http.HttpRequest;
+
+enum HttpHeaderSetter implements TextMapSetter<HttpRequest> {
+  INSTANCE;
+
+  @Override
+  public void set(@Nullable HttpRequest carrier, String key, String value) {
+    if (carrier == null) {
+      return;
+    }
+    carrier.setHeader(key, value);
+  }
+}
