+++ settings.gradle.kts
@@ -156,8 +156,7 @@ include(":instrumentation:apache-httpclient:apache-httpclient-4.0:javaagent")
 include(":instrumentation:apache-httpclient:apache-httpclient-4.3:library")
 include(":instrumentation:apache-httpclient:apache-httpclient-4.3:testing")
 include(":instrumentation:apache-httpclient:apache-httpclient-5.0:javaagent")
-include(":instrumentation:apache-httpclient:apache-httpclient-5.0:library")
-include(":instrumentation:apache-httpclient:apache-httpclient-5.0:testing")
+include(":instrumentation:apache-httpclient:apache-httpclient-5.2:library")
 include(":instrumentation:armeria-1.3:javaagent")
 include(":instrumentation:armeria-1.3:library")
 include(":instrumentation:armeria-1.3:testing")
